+++ akshare/__init__.py
@@ -2496,9 +2496,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.10.79 add: add stock_hold_management_detail_em interface
 1.10.80 fix: fix macro_china_gyzjz interface
 1.10.81 fix: fix stock_bid_ask_em interface
+1.10.82 fix: fix currency_boc_sina interface
 """
 
-__version__ = "1.10.81"
+__version__ = "1.10.82"
 __author__ = "AKFamily"
 
 import sys