+++ docs/changelog.md
@@ -63,6 +63,12 @@
 
 ## 更新说明详情
 
+1.10.82 fix: fix currency_boc_sina interface
+
+    1. 修复 currency_boc_sina 接口
+    2. 修复 futures_dce_position_rank 接口
+    3. 修复 get_rank_sum_daily 接口
+
 1.10.81 fix: fix stock_bid_ask_em interface
 
     1. 修复 stock_bid_ask_em 接口
@@ -2749,6 +2755,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.10.82 fix: fix currency_boc_sina interface
 
 1.10.81 fix: fix stock_bid_ask_em interface